feat(nuxt): unify loading api behind hooks
