+++ packages/nuxt/src/pages/runtime/plugins/router.ts
@@ -135,6 +135,7 @@ const plugin: Plugin<{ router: Router }> = defineNuxtPlugin({
 
     const initialLayout = nuxtApp.payload.state._layout
     router.beforeEach(async (to, from) => {
+      nuxtApp.callHook('page:loading:start')
       to.meta = reactive(to.meta)
       if (nuxtApp.isHydrating && initialLayout && !isReadonly(to.meta.layout)) {
         to.meta.layout = initialLayout as Exclude<PageMeta['layout'], Ref | false>
@@ -186,9 +187,15 @@ const plugin: Plugin<{ router: Router }> = defineNuxtPlugin({
       }
     })
 
-    router.onError(() => { delete nuxtApp._processingMiddleware })
+    router.onError(() => {
+      nuxtApp.callHook('page:loading:end')
+      delete nuxtApp._processingMiddleware
+    })
 
     router.afterEach(async (to, _from, failure) => {
+      if (failure) {
+        nuxtApp.callHook('page:loading:end')
+      }
       delete nuxtApp._processingMiddleware
 
       if (import.meta.client && !nuxtApp.isHydrating && error.value) {
